--- src/main/java/site/iris/issuefy/model/dto/TestDto.java
@@ -1,12 +1,0 @@
-package site.iris.issuefy.model.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class TestDto {
-	private long id;
-	private String message;
-
-}
